Moves some spring tests to be integration tests to break circle less
CircleCI tends to kill our process during server tests. This moves
some to be integration tests, which hopefully can keep it alive longer.
Travis still runs these tests.
